Move note creation onto SupportTaskService
Move CreateNoteAsync from SupportTaskNoteService onto SupportTaskService
and delete SupportTaskNoteService.
