--- src/TeachingRecordSystem.Core/Services/SupportTaskNotes/Extensions.cs
@@ -1,13 +1,0 @@
-using Microsoft.Extensions.DependencyInjection;
-
-namespace TeachingRecordSystem.Core.Services.SupportTaskNotes;
-
-public static class Extensions
-{
-    public static IServiceCollection AddSupportTaskNoteService(this IServiceCollection services)
-    {
-        services.AddTransient<SupportTaskNoteService>();
-
-        return services;
-    }
-}
